Combined all code. (see description)
I finished extracting and changing the old gui form the logic. 
only one instance of ListOfSites should be created and all the data should be stored in that structure. All the methods you should need to implement the functions of the GUI should be added. If I missed something or if any method does not make sense, I am happy to add new methods or explain how to use them.